code coverage increased to 100%

+++ esp32-mock/FS.h
@@ -1,3 +1,16 @@
+// Copyright 2024 Rik Essenius
+// 
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+// except in compliance with the License. You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and limitations under the License.
+
+// Mock implementation of the ESP library for unit testing (not targeting the ESP32)
+
 #ifndef FS_H
 #define FS_H
 #include "StringArduino.h"
@@ -27,11 +40,11 @@ public:
     File() = default;
     bool _exists = false;
     explicit operator bool() const;
-    int available() const;
+    size_t available() const;
     void close();
     size_t position() const;
     void seek(int offset, SeekMode mode);
-    int size() const;
+    size_t size() const;
     String readString();
     void write(const char* str, size_t length);
